fixup! CODENVY-581: Сache svn credentials
