+++ ide/che-core-ide-api/src/main/java/org/eclipse/che/ide/api/subversion/Credentials.java
@@ -19,6 +19,11 @@ public class Credentials {
     private String username;
     private String password;
 
+    public Credentials(String username, String password) {
+        this.username = username;
+        this.password = password;
+    }
+
     /**
      * Returns user name for authentication.
      */
