+++ plugins/plugin-svn/che-plugin-svn-ext-ide/src/main/java/org/eclipse/che/plugin/svn/ide/merge/MergePresenter.java
@@ -166,19 +166,15 @@ public class MergePresenter extends SubversionActionPresenter implements MergeVi
                    @Override
                    public void apply(PromiseError error) throws OperationException {
                        if (getErrorCode(error.getCause()) == UNAUTHORIZED_SVN_OPERATION) {
-                           notificationManager.notify(constants.authenticationFailed(), FAIL, FLOAT_MODE);
-
-                           subversionCredentialsDialog.askCredentials().then(new Operation<Credentials>() {
-                               @Override
-                               public void apply(Credentials credentials) throws OperationException {
-                                   merge(project, resources, credentials);
-                               }
-                           }).catchError(new Operation<PromiseError>() {
-                               @Override
-                               public void apply(PromiseError error) throws OperationException {
-                                   notificationManager.notify(error.getMessage(), FAIL, FLOAT_MODE);
-                               }
-                           });
+                           tryWithCredentials(notificationManager,
+                                              subversionCredentialsDialog,
+                                              constants.authenticationFailed(),
+                                              new SVNOperation() {
+                                                  @Override
+                                                  public void perform(Credentials credentials) {
+                                                      merge(project, resources, credentials);
+                                                  }
+                                              });
                        } else {
                            notificationManager.notify(error.getMessage(), FAIL, FLOAT_MODE);
                        }
